Clean up headers
