+++ include/mls/crypto.h
@@ -1,11 +1,66 @@
 #pragma once
 
-#include "mls/common.h"
+#include <mls/common.h>
 #include <openssl/evp.h>
 #include <tls/tls_syntax.h>
+#include <hpke/digest.h>
+#include <hpke/hpke.h>
+#include <hpke/signature.h>
+#include <hpke/random.h>
 #include <vector>
 
 namespace mls {
+
+/// Cipher suites
+struct CipherSuite
+{
+  enum struct ID : uint16_t
+  {
+    unknown = 0x0000,
+    X25519_AES128GCM_SHA256_Ed25519 = 0x0001,
+    P256_AES128GCM_SHA256_P256 = 0x0002,
+    X25519_CHACHA20POLY1305_SHA256_Ed25519 = 0x0003,
+    X448_AES256GCM_SHA512_Ed448 = 0x0004,
+    P521_AES256GCM_SHA512_P521 = 0x0005,
+    X448_CHACHA20POLY1305_SHA512_Ed448 = 0x0006,
+  };
+
+  CipherSuite();
+  CipherSuite(ID id_in);
+  CipherSuite(const CipherSuite& other);
+  CipherSuite(CipherSuite&& other);
+  CipherSuite& operator=(const CipherSuite& other);
+
+  ID id;
+  std::unique_ptr<hpke::HPKE> hpke;
+  std::unique_ptr<hpke::Digest> digest;
+  std::unique_ptr<hpke::Signature> sig;
+
+  bytes expand_with_label(const bytes& secret,
+                          const std::string& label,
+                          const bytes& context,
+                          size_t size) const;
+
+private:
+  void reset(ID id_in);
+};
+
+tls::istream&
+operator>>(tls::istream& str, CipherSuite& suite);
+
+tls::ostream&
+operator<<(tls::ostream& str, const CipherSuite& suite);
+
+bool
+operator==(const CipherSuite& lhs, const CipherSuite& rhs);
+
+bool
+operator!=(const CipherSuite& lhs, const CipherSuite& rhs);
+
+extern const std::array<CipherSuite::ID, 6> all_supported_suites;
+
+// Utilities
+using hpke::random_bytes;
 
 bool
 constant_time_eq(const bytes& lhs, const bytes& rhs);
