+++ test/state.cpp
@@ -11,7 +11,7 @@ public:
   StateTest()
   {
     for (size_t i = 0; i < group_size; i += 1) {
-      auto init_secret = hpke::random_bytes(32);
+      auto init_secret = random_bytes(32);
       auto identity_priv = SignaturePrivateKey::generate(suite);
       auto credential = Credential::basic(user_id, identity_priv.public_key);
       auto init_priv = HPKEPrivateKey::derive(suite, init_secret);
@@ -40,7 +40,7 @@ protected:
 
   bytes fresh_secret() const
   {
-    return hpke::random_bytes(suite.hpke->kdf->hash_size());
+    return random_bytes(suite.hpke->kdf->hash_size());
   }
 };
